Grids: move filter modules into per-module directories
