+++ packages/devextreme/js/__internal/grids/data_grid/module_not_extended/filter_panel.ts
@@ -1,4 +1,4 @@
-import { filterPanelModule } from '@ts/grids/grid_core/filter/m_filter_panel';
+import { filterPanelModule } from '@ts/grids/grid_core/filter_panel/m_filter_panel';
 
 import core from '../m_core';
