+++ packages/devextreme/js/__internal/grids/data_grid/module_not_extended/filter_row.ts
@@ -1,4 +1,4 @@
-import { filterRowModule } from '@ts/grids/grid_core/filter/m_filter_row';
+import { filterRowModule } from '@ts/grids/grid_core/filter_row/m_filter_row';
 
 import core from '../m_core';
