+++ packages/devextreme/js/__internal/grids/data_grid/module_not_extended/filter_sync.ts
@@ -1,4 +1,4 @@
-import { filterSyncModule } from '@ts/grids/grid_core/filter/m_filter_sync';
+import { filterSyncModule } from '@ts/grids/grid_core/filter_sync/m_filter_sync';
 
 import core from '../m_core';
