+++ packages/devextreme/js/__internal/grids/grid_core/filter_sync/m_filter_custom_operations.ts
