+++ packages/devextreme/js/__internal/grids/new/grid_core/filtering/filter_panel/view.tsx
@@ -1,6 +1,6 @@
 import { computed, type ReadonlySignal } from '@ts/core/state_manager/index';
-import { FilterBuilderView as OldFilterBuilderView } from '@ts/grids/grid_core/filter/m_filter_builder';
-import { FilterPanelView as OldFilterPanelView } from '@ts/grids/grid_core/filter/m_filter_panel';
+import { FilterBuilderView as OldFilterBuilderView } from '@ts/grids/grid_core/filter_builder/m_filter_builder';
+import { FilterPanelView as OldFilterPanelView } from '@ts/grids/grid_core/filter_panel/m_filter_panel';
 
 import { View } from '../../core/view';
 import { WidgetMock } from '../../widget_mock';
