+++ packages/devextreme/js/__internal/grids/tree_list/module_not_extended/filter_builder.ts
@@ -1,4 +1,4 @@
-import { filterBuilderModule } from '@ts/grids/grid_core/filter/m_filter_builder';
+import { filterBuilderModule } from '@ts/grids/grid_core/filter_builder/m_filter_builder';
 
 import core from '../m_core';
